requirements.txt
https://download.pytorch.org/whl/cpu/torch-1.3.1%2Bcpu-cp37-cp37m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.4.2%2Bcpu-cp37-cp37m-linux_x86_64.whl
ipywidgets>=7.5.1
#voila>=0.2.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cycler==0.11.0
Flask==2.2.5
Flask-WTF==1.1.1
fonttools==4.38.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.8.0.76
packaging==23.1
Pillow==9.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
